requirements.txt
# =====================================================
# eva_ontology_graphrag_demo1 - Gemini API 버전
# Python 3.12
# =====================================================

# Core
streamlit>=1.32.0
pyyaml>=6.0

# LLM - Gemini API
google-generativeai>=0.8.0

# 환경변수 로드 (로컬 개발용)
python-dotenv>=1.0.0

# Vector DB
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0

# Graph
networkx>=3.2
neo4j>=5.18.0

# Document parsing
python-docx>=1.1.0
pypdf>=4.1.0
olefile>=0.47

# ML
torch>=2.2.0
numpy>=1.26.0

# Utils
tqdm>=4.66.0